Map everything to the config resolver. Separate parsers to a specific module.
